backend/requirements.txt
# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Clerk Integration
svix==1.15.0
clerk-sdk-python==0.1.1

# Python Utilities
python-dotenv==1.0.0
pytz==2023.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
